requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Vector database and search
faiss-cpu>=1.7.4
chromadb>=0.4.18

# Data processing
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0

# LLM inference
huggingface-hub>=0.19.0

# Evaluation and metrics
scikit-learn>=1.3.0
nltk>=3.8.0
rouge-score>=0.1.2

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Jupyter for experiments
jupyter>=1.0.0
ipywidgets>=8.1.0
